Migrate to cipher v0.5.0-pre.7
