+++ Cargo.toml
@@ -25,3 +25,6 @@ members = [
 
 [profile.dev]
 opt-level = 2
+
+[patch.crates-io]
+cipher = { git = "https://github.com/RustCrypto/traits", branch = "block_backends" }
