+++ belt-block/tests/mod.rs
@@ -3,7 +3,10 @@
 use belt_block::{belt_block_raw, belt_wblock_dec, belt_wblock_enc};
 #[cfg(feature = "cipher")]
 use belt_block::{
-    cipher::{BlockCipherDecrypt, BlockCipherEncrypt, KeyInit},
+    cipher::{
+        block::{BlockCipherDecrypt, BlockCipherEncrypt},
+        KeyInit,
+    },
     BeltBlock,
 };
 use hex_literal::hex;
