+++ cast6/tests/mod.rs
@@ -1,5 +1,8 @@
 use cast6::Cast6;
-use cipher::{Block, BlockCipherDecrypt, BlockCipherEncrypt, KeyInit};
+use cipher::{
+    block::{BlockCipherDecrypt, BlockCipherEncrypt},
+    Block, KeyInit,
+};
 use hex_literal::hex;
 
 /// Test vectors from RFC 2612 Appendix A
