+++ des/src/tdes.rs
@@ -1,9 +1,13 @@
 //! Triple DES (3DES) block ciphers.
 
-use crate::des::{gen_keys, Des};
+use crate::{utils::gen_keys, Des};
 use cipher::{
-    consts::{U16, U24, U8},
-    AlgorithmName, BlockCipher, Key, KeyInit, KeySizeUser,
+    block::{
+        BlockCipherDecBackend, BlockCipherDecClosure, BlockCipherDecrypt, BlockCipherEncBackend,
+        BlockCipherEncClosure, BlockCipherEncrypt,
+    },
+    consts::{U1, U16, U24, U8},
+    AlgorithmName, Block, BlockSizeUser, InOut, Key, KeyInit, KeySizeUser, ParBlocksSizeUser,
 };
 use core::fmt;
 
@@ -17,8 +21,6 @@ pub struct TdesEde3 {
     d2: Des,
     d3: Des,
 }
-
-impl BlockCipher for TdesEde3 {}
 
 impl KeySizeUser for TdesEde3 {
     type KeySize = U24;
@@ -37,6 +39,50 @@ impl KeyInit for TdesEde3 {
     }
 }
 
+impl BlockSizeUser for TdesEde3 {
+    type BlockSize = U8;
+}
+
+impl ParBlocksSizeUser for TdesEde3 {
+    type ParBlocksSize = U1;
+}
+
+impl BlockCipherEncrypt for TdesEde3 {
+    #[inline]
+    fn encrypt_with_backend(&self, f: impl BlockCipherEncClosure<BlockSize = Self::BlockSize>) {
+        f.call(self)
+    }
+}
+
+impl BlockCipherEncBackend for TdesEde3 {
+    #[inline]
+    fn encrypt_block(&self, mut block: InOut<'_, '_, Block<Self>>) {
+        let mut data = u64::from_be_bytes(block.clone_in().into());
+        data = self.d1.encrypt(data);
+        data = self.d2.decrypt(data);
+        data = self.d3.encrypt(data);
+        block.get_out().copy_from_slice(&data.to_be_bytes());
+    }
+}
+
+impl BlockCipherDecrypt for TdesEde3 {
+    #[inline]
+    fn decrypt_with_backend(&self, f: impl BlockCipherDecClosure<BlockSize = Self::BlockSize>) {
+        f.call(self)
+    }
+}
+
+impl BlockCipherDecBackend for TdesEde3 {
+    #[inline]
+    fn decrypt_block(&self, mut block: InOut<'_, '_, Block<Self>>) {
+        let mut data = u64::from_be_bytes(block.clone_in().into());
+        data = self.d3.decrypt(data);
+        data = self.d2.encrypt(data);
+        data = self.d1.decrypt(data);
+        block.get_out().copy_from_slice(&data.to_be_bytes());
+    }
+}
+
 impl fmt::Debug for TdesEde3 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.write_str("TdesEee3 { ... }")
@@ -50,26 +96,7 @@ impl AlgorithmName for TdesEde3 {
 }
 
 #[cfg(feature = "zeroize")]
-#[cfg_attr(docsrs, doc(cfg(feature = "zeroize")))]
 impl ZeroizeOnDrop for TdesEde3 {}
-
-cipher::impl_simple_block_encdec!(
-    TdesEde3, U8, cipher, block,
-    encrypt: {
-        let mut data = u64::from_be_bytes(block.clone_in().into());
-        data = cipher.d1.encrypt(data);
-        data = cipher.d2.decrypt(data);
-        data = cipher.d3.encrypt(data);
-        block.get_out().copy_from_slice(&data.to_be_bytes());
-    }
-    decrypt: {
-        let mut data = u64::from_be_bytes(block.clone_in().into());
-        data = cipher.d3.decrypt(data);
-        data = cipher.d2.encrypt(data);
-        data = cipher.d1.decrypt(data);
-        block.get_out().copy_from_slice(&data.to_be_bytes());
-    }
-);
 
 /// Triple DES (3DES) block cipher.
 #[derive(Clone)]
@@ -78,8 +105,6 @@ pub struct TdesEee3 {
     d2: Des,
     d3: Des,
 }
-
-impl BlockCipher for TdesEee3 {}
 
 impl KeySizeUser for TdesEee3 {
     type KeySize = U24;
@@ -98,6 +123,50 @@ impl KeyInit for TdesEee3 {
     }
 }
 
+impl BlockSizeUser for TdesEee3 {
+    type BlockSize = U8;
+}
+
+impl ParBlocksSizeUser for TdesEee3 {
+    type ParBlocksSize = U1;
+}
+
+impl BlockCipherEncrypt for TdesEee3 {
+    #[inline]
+    fn encrypt_with_backend(&self, f: impl BlockCipherEncClosure<BlockSize = Self::BlockSize>) {
+        f.call(self)
+    }
+}
+
+impl BlockCipherEncBackend for TdesEee3 {
+    #[inline]
+    fn encrypt_block(&self, mut block: InOut<'_, '_, Block<Self>>) {
+        let mut data = u64::from_be_bytes(block.clone_in().into());
+        data = self.d1.encrypt(data);
+        data = self.d2.encrypt(data);
+        data = self.d3.encrypt(data);
+        block.get_out().copy_from_slice(&data.to_be_bytes());
+    }
+}
+
+impl BlockCipherDecrypt for TdesEee3 {
+    #[inline]
+    fn decrypt_with_backend(&self, f: impl BlockCipherDecClosure<BlockSize = Self::BlockSize>) {
+        f.call(self)
+    }
+}
+
+impl BlockCipherDecBackend for TdesEee3 {
+    #[inline]
+    fn decrypt_block(&self, mut block: InOut<'_, '_, Block<Self>>) {
+        let mut data = u64::from_be_bytes(block.clone_in().into());
+        data = self.d3.decrypt(data);
+        data = self.d2.decrypt(data);
+        data = self.d1.decrypt(data);
+        block.get_out().copy_from_slice(&data.to_be_bytes());
+    }
+}
+
 impl fmt::Debug for TdesEee3 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.write_str("TdesEee3 { ... }")
@@ -111,26 +180,7 @@ impl AlgorithmName for TdesEee3 {
 }
 
 #[cfg(feature = "zeroize")]
-#[cfg_attr(docsrs, doc(cfg(feature = "zeroize")))]
 impl ZeroizeOnDrop for TdesEee3 {}
-
-cipher::impl_simple_block_encdec!(
-    TdesEee3, U8, cipher, block,
-    encrypt: {
-        let mut data = u64::from_be_bytes(block.clone_in().into());
-        data = cipher.d1.encrypt(data);
-        data = cipher.d2.encrypt(data);
-        data = cipher.d3.encrypt(data);
-        block.get_out().copy_from_slice(&data.to_be_bytes());
-    }
-    decrypt: {
-        let mut data = u64::from_be_bytes(block.clone_in().into());
-        data = cipher.d3.decrypt(data);
-        data = cipher.d2.decrypt(data);
-        data = cipher.d1.decrypt(data);
-        block.get_out().copy_from_slice(&data.to_be_bytes());
-    }
-);
 
 /// Triple DES (3DES) block cipher.
 #[derive(Clone)]
@@ -138,8 +188,6 @@ pub struct TdesEde2 {
     d1: Des,
     d2: Des,
 }
-
-impl BlockCipher for TdesEde2 {}
 
 impl KeySizeUser for TdesEde2 {
     type KeySize = U16;
@@ -156,6 +204,50 @@ impl KeyInit for TdesEde2 {
     }
 }
 
+impl BlockSizeUser for TdesEde2 {
+    type BlockSize = U8;
+}
+
+impl ParBlocksSizeUser for TdesEde2 {
+    type ParBlocksSize = U1;
+}
+
+impl BlockCipherEncrypt for TdesEde2 {
+    #[inline]
+    fn encrypt_with_backend(&self, f: impl BlockCipherEncClosure<BlockSize = Self::BlockSize>) {
+        f.call(self)
+    }
+}
+
+impl BlockCipherEncBackend for TdesEde2 {
+    #[inline]
+    fn encrypt_block(&self, mut block: InOut<'_, '_, Block<Self>>) {
+        let mut data = u64::from_be_bytes(block.clone_in().into());
+        data = self.d1.encrypt(data);
+        data = self.d2.decrypt(data);
+        data = self.d1.encrypt(data);
+        block.get_out().copy_from_slice(&data.to_be_bytes());
+    }
+}
+
+impl BlockCipherDecrypt for TdesEde2 {
+    #[inline]
+    fn decrypt_with_backend(&self, f: impl BlockCipherDecClosure<BlockSize = Self::BlockSize>) {
+        f.call(self)
+    }
+}
+
+impl BlockCipherDecBackend for TdesEde2 {
+    #[inline]
+    fn decrypt_block(&self, mut block: InOut<'_, '_, Block<Self>>) {
+        let mut data = u64::from_be_bytes(block.clone_in().into());
+        data = self.d1.decrypt(data);
+        data = self.d2.encrypt(data);
+        data = self.d1.decrypt(data);
+        block.get_out().copy_from_slice(&data.to_be_bytes());
+    }
+}
+
 impl fmt::Debug for TdesEde2 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.write_str("TdesEde2 { ... }")
@@ -169,26 +261,7 @@ impl AlgorithmName for TdesEde2 {
 }
 
 #[cfg(feature = "zeroize")]
-#[cfg_attr(docsrs, doc(cfg(feature = "zeroize")))]
 impl ZeroizeOnDrop for TdesEde2 {}
-
-cipher::impl_simple_block_encdec!(
-    TdesEde2, U8, cipher, block,
-    encrypt: {
-        let mut data = u64::from_be_bytes(block.clone_in().into());
-        data = cipher.d1.encrypt(data);
-        data = cipher.d2.decrypt(data);
-        data = cipher.d1.encrypt(data);
-        block.get_out().copy_from_slice(&data.to_be_bytes());
-    }
-    decrypt: {
-        let mut data = u64::from_be_bytes(block.clone_in().into());
-        data = cipher.d1.decrypt(data);
-        data = cipher.d2.encrypt(data);
-        data = cipher.d1.decrypt(data);
-        block.get_out().copy_from_slice(&data.to_be_bytes());
-    }
-);
 
 /// Triple DES (3DES) block cipher.
 #[derive(Clone)]
@@ -196,8 +269,6 @@ pub struct TdesEee2 {
     d1: Des,
     d2: Des,
 }
-
-impl BlockCipher for TdesEee2 {}
 
 impl KeySizeUser for TdesEee2 {
     type KeySize = U16;
@@ -214,6 +285,50 @@ impl KeyInit for TdesEee2 {
     }
 }
 
+impl BlockSizeUser for TdesEee2 {
+    type BlockSize = U8;
+}
+
+impl ParBlocksSizeUser for TdesEee2 {
+    type ParBlocksSize = U1;
+}
+
+impl BlockCipherEncrypt for TdesEee2 {
+    #[inline]
+    fn encrypt_with_backend(&self, f: impl BlockCipherEncClosure<BlockSize = Self::BlockSize>) {
+        f.call(self)
+    }
+}
+
+impl BlockCipherEncBackend for TdesEee2 {
+    #[inline]
+    fn encrypt_block(&self, mut block: InOut<'_, '_, Block<Self>>) {
+        let mut data = u64::from_be_bytes(block.clone_in().into());
+        data = self.d1.encrypt(data);
+        data = self.d2.encrypt(data);
+        data = self.d1.encrypt(data);
+        block.get_out().copy_from_slice(&data.to_be_bytes());
+    }
+}
+
+impl BlockCipherDecrypt for TdesEee2 {
+    #[inline]
+    fn decrypt_with_backend(&self, f: impl BlockCipherDecClosure<BlockSize = Self::BlockSize>) {
+        f.call(self)
+    }
+}
+
+impl BlockCipherDecBackend for TdesEee2 {
+    #[inline]
+    fn decrypt_block(&self, mut block: InOut<'_, '_, Block<Self>>) {
+        let mut data = u64::from_be_bytes(block.clone_in().into());
+        data = self.d1.decrypt(data);
+        data = self.d2.decrypt(data);
+        data = self.d1.decrypt(data);
+        block.get_out().copy_from_slice(&data.to_be_bytes());
+    }
+}
+
 impl fmt::Debug for TdesEee2 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.write_str("TdesEee2 { ... }")
@@ -227,23 +342,4 @@ impl AlgorithmName for TdesEee2 {
 }
 
 #[cfg(feature = "zeroize")]
-#[cfg_attr(docsrs, doc(cfg(feature = "zeroize")))]
 impl ZeroizeOnDrop for TdesEee2 {}
-
-cipher::impl_simple_block_encdec!(
-    TdesEee2, U8, cipher, block,
-    encrypt: {
-        let mut data = u64::from_be_bytes(block.clone_in().into());
-        data = cipher.d1.encrypt(data);
-        data = cipher.d2.encrypt(data);
-        data = cipher.d1.encrypt(data);
-        block.get_out().copy_from_slice(&data.to_be_bytes());
-    }
-    decrypt: {
-        let mut data = u64::from_be_bytes(block.clone_in().into());
-        data = cipher.d1.decrypt(data);
-        data = cipher.d2.decrypt(data);
-        data = cipher.d1.decrypt(data);
-        block.get_out().copy_from_slice(&data.to_be_bytes());
-    }
-);
